requirements.txt
# Core ML Libraries
ultralytics==8.0.220
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
pyyaml==6.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Experiment Tracking
wandb==0.16.1

# Web Application
gradio==4.8.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Jupyter (optional, for notebooks)
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# Model Export
onnx==1.15.0
onnxruntime==1.16.3

# Cloud Integration
huggingface-hub==0.19.4
kaggle==1.5.16

# Development
ipdb==0.13.13
pre-commit==3.5.0